# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sabi Health Backend API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: Canonical frontend and operations reference generated from the registered Express routes and validation contracts. All currency values ending in Minor are NGN minor units (kobo).
servers:
  - url: /
    description: Current environment
...
paths:
...
  /api/v1/notifications:
    get:
      tags:
        - Notifications
      summary: List notifications
      operationId: get__api_v1_notifications
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/v1/notifications/settings:
    get:
      tags:
        - Notifications
      summary: My notification settings
      description: WhatsApp availability, the linked number (masked), a code being verified, the kinds of updates sent to WhatsApp and the consent on record. The in-app bell always receives everything.
      operationId: get__api_v1_notifications_settings
      security:
        - bearerAuth: []
      responses:
        "200": &a1
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags:
        - Notifications
      summary: Change notification settings
      description: Choose which kinds of update go to WhatsApp, whether reminders may name the medicine (off by default; reminders then say "your morning medicine"), and the time zone used for new schedules.
      operationId: put__api_v1_notifications_settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                whatsappCategories:
                  type: array
                  items:
                    type: string
                    enum:
                      - MEDICATION
                      - APPOINTMENT
                      - CARE
                showMedicationDetails:
                  type: boolean
                timezone:
                  type: string
                  example: Africa/Lagos
      responses:
        "200": *a1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/notifications/settings/whatsapp:
    post:
      tags:
        - Notifications
      summary: Send a WhatsApp verification code
      description: "Step 1 of enabling WhatsApp or changing the number: the patient agrees to WhatsApp notifications and Sabi sends a 6-digit code by WhatsApp authentication template. Codes expire after 10 minutes; at most 5 per hour. A number linked to another account is refused (409 NUMBER_IN_USE). The current number keeps working until the new one is verified."
      operationId: post__api_v1_notifications_settings_whatsapp
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - consent
              additionalProperties: false
              properties:
                phone:
                  type: string
                  example: 0803 123 4567
                consent:
                  const: true
      responses:
        "200": *a1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags:
        - Notifications
      summary: Turn off WhatsApp notifications
      description: Unlinks the number and cancels WhatsApp messages still waiting to be sent. In-app notifications are unaffected.
      operationId: delete__api_v1_notifications_settings_whatsapp
      security:
        - bearerAuth: []
      responses:
        "200": *a1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/notifications/settings/whatsapp/resend:
    post:
      tags:
        - Notifications
      summary: Send the code again
      description: Available 60 seconds after the previous code.
      operationId: post__api_v1_notifications_settings_whatsapp_resend
      security:
        - bearerAuth: []
      responses:
        "200": *a1
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/notifications/settings/whatsapp/verify:
    post:
      tags:
        - Notifications
      summary: Verify the WhatsApp code
      description: "Step 2: the number becomes the patient's WhatsApp number, any previous number is revoked, and the consent version and time are recorded. Five wrong codes end the attempt."
      operationId: post__api_v1_notifications_settings_whatsapp_verify
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              additionalProperties: false
              properties:
                code:
                  type: string
                  pattern: ^\d{6}$
      responses:
        "200": *a1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Use an access token. Example: Bearer <access-token>"
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        status:
          const: error
        message:
          type: string
        code:
          type: string
      required:
        - status
        - message
    Success:
      type: object
      additionalProperties: false
      properties:
        status:
          const: success
        data: {}
      required:
        - status
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource is absent or hidden by authorization scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: State conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Safe module-scoped server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...